+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2025-04-10
+
+### Changed
+
+- switched out the optional dep on `axelson/publicsuffix-elixir` for [public_sufx](https://hex.pm/packages/public_sufx), a fork that's published on hex.pm
+- removed the `:reject_public_suffixes` option, public suffixes are no always rejected to guard against 'supercookies'
+
 ## [0.7.0] - 2025-02-17
 
 ### Added